+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `http` and `grpc` `ServerOptions` to wrap a slice of options into a single element
 - Business error middleware
+- Custom endpoint middleware chain
+
 
 ## 0.1.0 - 2019-09-25